Shift checkIfAbort to InputParser
