+++ src/main/java/seedu/mtracker/console/AddCryptoParser.java
@@ -7,8 +7,6 @@ import seedu.mtracker.commons.Validate;
 import seedu.mtracker.error.OperationAbortedError;
 import seedu.mtracker.ui.TextUi;
 
-import static seedu.mtracker.commons.Validate.checkIfAbort;
-
 public class AddCryptoParser extends AddInstrumentParser {
 
     public static String INSTRUMENT_TYPE = "crypto";
